requirements.txt
python-dotenv
scikit-learn
pandas